Added my-profile test & refactored
