--- Car/__tests__/activity/Preferences-test.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-import Preferences from "../../src/activity/my-profile/my-profile-activity/preferences/Preferences";
-
-test("renders correctly", () =>
-  expect(<Preferences />).toBeDefined());
