+++ Car/__tests__/activity/login/Login-test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Login from "../../src/activity/login/Login";
+import Login from "../../../src/activity/login/Login";
 
 test("renders correctly", () =>
   expect(<Login />).toBeDefined());
