+++ Car/__tests__/activity/my-profile/AddCars-test.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+import AddCars from "../../../src/activity/my-profile/my-profile-activity/cars/car-activity/add-cars/AddCars";
+
+test("renders correctly", () =>
+  expect(<AddCars />).toBeDefined());
