+++ Car/__tests__/activity/my-profile/AddressBook-test.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+import AddressBook from "../../../src/activity/my-profile/my-profile-activity/address-book/AddressBook";
+
+test("renders correctly", () =>
+  expect(<AddressBook />).toBeDefined());
