+++ Car/__tests__/activity/my-profile/Cars-test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Cars from "../../src/activity/my-profile/my-profile-activity/cars/Cars";
+import Cars from "../../../src/activity/my-profile/my-profile-activity/cars/Cars";
 
 test("renders correctly", () =>
   expect(<Cars />).toBeDefined());
