+++ Car/__tests__/activity/my-profile/Details-test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Details from "../../src/activity/my-profile/my-profile-activity/details/Details";
+import Details from "../../../src/activity/my-profile/my-profile-activity/details/Details";
 
 test("renders correctly", () =>
   expect(<Details />).toBeDefined());
