+++ Car/__tests__/activity/my-profile/EditCars-test.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+import EditCars from "../../../src/activity/my-profile/my-profile-activity/cars/car-activity/edit-cars/EditCars";
+
+test("renders correctly", () =>
+  expect(<EditCars />).toBeDefined());
