+++ Car/__tests__/activity/my-profile/Preferences-test.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+import Preferences from "../../../src/activity/my-profile/my-profile-activity/preferences/Preferences";
+
+test("renders correctly", () =>
+  expect(<Preferences />).toBeDefined());
