+++ Car/__tests__/activity/my-profile/SetPlace-test.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+import SetPlace from "../../../src/activity/my-profile/my-profile-activity/address-book/address-book-activity/add-locations/SetPlace";
+
+test("renders correctly", () =>
+  expect(<SetPlace />).toBeDefined());
