+++ Car/__tests__/activity/my-profile/Settings-test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Settings from "../../src/activity/my-profile/my-profile-activity/settings/Settings";
+import Settings from "../../../src/activity/my-profile/my-profile-activity/settings/Settings";
 
 test("renders correctly", () =>
   expect(<Settings />).toBeDefined());
